+++ docs/scanners/yarn_audit.md
@@ -1,0 +1,5 @@
+# [YarnAudit](https://yarnpkg.com/lang/en/docs/cli/audit/)
+
+Finds CVEs in Node modules included as dependencies in a project that is packaged by Yarn.
+
+See the [NodeAudit](/docs/node_audit.md) doc for configuration options.
