+++ spec/fixtures/config/node_audit_config.yaml
@@ -1,0 +1,3 @@
+scanner_configs:
+  NodeAudit:
+    foo: bar
