+++ spec/fixtures/config/npm_audit_config.yaml
@@ -1,0 +1,6 @@
+scanner_configs:
+  NPMAudit:
+    exceptions:
+      - advisory_id: '12'
+        changed_by: appsec team
+        notes: barfoo
