+++ spec/fixtures/yarn_audit/success_with_exceptions/salus.yaml
@@ -1,0 +1,13 @@
+scanner_configs:
+    YarnAudit:
+        exceptions:
+            - {
+                advisory_id: "39",
+                changed_by: "ryan.sears",
+                notes: "See https://www.npmjs.com/advisories/39. We're not vulnerable to this because uglify is a transient dev dependency, and we never actually invoke it with our code.",
+              }
+            - {
+                advisory_id: "48",
+                changed_by: "ryan.sears",
+                notes: "See https://www.npmjs.com/advisories/48. We're not vulnerable to this because this is a regex dos and we have nothing that puts user input into it. The impact is also minimal.",
+              }
