+++ spec/lib/salus/scanners/yarn_audit_spec.rb
@@ -1,20 +1,20 @@
 require_relative '../../../spec_helper.rb'
 
-describe Salus::Scanners::NPMAudit do
+describe Salus::Scanners::YarnAudit do
   describe '#should_run?' do
     it 'should return false in the absence of package.json and friends' do
       repo = Salus::Repo.new('spec/fixtures/blank_repository')
-      expect(repo.package_lock_json_present?).to eq(false)
+      expect(repo.yarn_lock_present?).to eq(false)
 
-      scanner = Salus::Scanners::NPMAudit.new(repository: repo, config: {})
+      scanner = Salus::Scanners::YarnAudit.new(repository: repo, config: {})
       expect(scanner.should_run?).to eq(false)
     end
 
     it 'should return true if package.json is present' do
-      repo = Salus::Repo.new('spec/fixtures/npm_audit/success')
-      expect(repo.package_lock_json_present?).to eq(true)
+      repo = Salus::Repo.new('spec/fixtures/yarn_audit/success')
+      expect(repo.yarn_lock_present?).to eq(true)
 
-      scanner = Salus::Scanners::NPMAudit.new(repository: repo, config: {})
+      scanner = Salus::Scanners::YarnAudit.new(repository: repo, config: {})
       expect(scanner.should_run?).to eq(true)
     end
   end
