+++ yarn-error.log
@@ -1,0 +1,41 @@
+Arguments: 
+  /usr/local/Cellar/node/10.4.1/bin/node /usr/local/bin/yarn audit
+
+PATH: 
+  /Users/jborrey/.rvm/gems/ruby-2.4.4/bin:/Users/jborrey/.rvm/gems/ruby-2.4.4@global/bin:/Users/jborrey/.rvm/rubies/ruby-2.4.4/bin:/Users/jborrey/.rvm/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/go/bin:/usr/local/MacGPG2/bin
+
+Yarn version: 
+  1.12.1
+
+Node version: 
+  10.4.1
+
+Platform: 
+  darwin x64
+
+Trace: 
+  Error: Unexpected audit response (Missing Metadata): false
+      at Audit.<anonymous> (/Users/jborrey/.config/yarn/global/node_modules/yarn/lib/cli.js:55104:15)
+      at Generator.next (<anonymous>)
+      at step (/Users/jborrey/.config/yarn/global/node_modules/yarn/lib/cli.js:92:30)
+      at /Users/jborrey/.config/yarn/global/node_modules/yarn/lib/cli.js:103:13
+      at process._tickCallback (internal/process/next_tick.js:68:7)
+
+npm manifest: 
+  {
+    "dependencies": {
+      "@yarnpkg/lockfile": "^1.0.1"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
+  # yarn lockfile v1
+  
+  
+  "@yarnpkg/lockfile@^1.0.1":
+    version "1.0.1"
+    resolved "https://registry.yarnpkg.com/@yarnpkg/lockfile/-/lockfile-1.0.1.tgz#294155ffbf78cec882a23bc61800b8e629b6b384"
